+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.15.0 (unreleased)
+*******************
+
+* Address distutils deprecation warning in Python 3.10 (:pr:`242`).
+  Thanks :user:`GabrielLins64` for the PR.
+
 0.14.0 (2020-09-27)
 *******************